- Updated `beginPos` parameter of GetLogEntries() to UInt32.
- Updated XML comments for GetLogEntries().
- Updated code style stuff.

NOTE: The log entry's has to be loaded progressively from the end to start of the log file. The `beginPos` parameter is used in combination of the `last_pos` response to retrieve more then 100 log entries. `last_pos` is a part of each response from the server after usage from GetLogEntries().

TODO:
- Re-Implement `loglevel` parameter for example `logadd loglevel={1-4} logmsg={text}`...